+++ dbentry/site/views/actions.py
@@ -1,5 +1,9 @@
+from django.contrib import messages
+from mizdb_watchlist.manager import get_manager
+
 from dbentry.actions.views import MergeView
 from dbentry.site.views.delete import DeleteSelectedView
+from dbentry.site.views.export import ExportActionView, has_export_permission
 from dbentry.utils import permission as perms
 from dbentry.utils.permission import has_delete_permission
 
@@ -51,3 +55,21 @@ def has_merge_permission(user, opts):
 )
 def merge_records(view, request, queryset):
     return MergeView.as_view(view=view, queryset=queryset)(request)
+
+
+@action(label="Merkliste", description="Ausgewählte Objekte zur Merkliste hinzufügen")
+def watchlist(view, request, queryset):
+    manager = get_manager(request)
+    manager.bulk_add(queryset)
+    messages.add_message(request, level=messages.SUCCESS, message="Erfolgreich zu meiner Merkliste hinzugefügt.")
+    return None
+
+
+@action(
+    permission_func=has_export_permission,  # TODO: set to has_view_permission to enable export action for all users
+    label="Exportieren",
+    description="Die ausgewählten Objekte exportieren",
+)
+def export(view, request, queryset):
+    view = ExportActionView.as_view(model=queryset.model, queryset=queryset, resource_class=view.resource_class)
+    return view(request)