+++ tests/settings.py
@@ -33,6 +33,8 @@ ALWAYS_INSTALLED_APPS = [
     "django_bootstrap5",
     "mizdb_inlines",
     "mizdb_tomselect",
+    "mizdb_watchlist",
+    "import_export",
 ]
 
 TEST_APPS = [