+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.16] - 2024-05-12
+
+### Added
+
+- watchlist feature for keeping user lists of model objects
+- enable exporting records
+- add "Monat" column to Ausgabe autocomplete
+
+### Changed
+
+- disallow empty URLs
+- allow anonymous users to use changelist selection
+ 
 ## [0.15] - 2024-04-09
 
 ### Changed